requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
Django==3.0
django-bootstrap4==2.3.1
django-ckeditor==6.0.0
django-contrib-comments==2.1.0
django-crispy-forms==1.11.1
django-filter==2.4.0
django-js-asset==1.2.2
django-mssql-backend==2.8.1
djangorestframework==3.12.2
docutils==0.16
importlib-metadata==2.1.1
Pillow==8.1.2
pyodbc==4.0.30
pytz==2021.1
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
zipp==3.4.1
